Parse Cardano Staking Rewards

Reads generic CSV outputs from pooltool.io and transposes these into a format which can be read by Portfolio Performance.

Rewards are classified as "Delivery (Inbound)" (German: "Einlieferung") which is a category that is different from "buy" and "interest".

parse_cardano(
  data,
  pp_security_name = "Cardano",
  pp_lang = "DE",
  securities_account = NULL,
  filename = NULL,
  dec = NULL
)

Arguments

data
character
The (generic) input CSV from https://pooltool.io.

pp_security_name
character
The name of the security in Portfolio Performance (e.g. "Kusama").

pp_lang
character
The currency of Portfolio Performance. Only German ("DE") and English ("EN") are supported.

securities_account
character
The name of the account to which to associate the deliveries. In German called "Depot".

filename
character
The output filename.

dec
character
Decimal separator. If not set, inferred by the value of argument pp_lang:

If pp_lang = "DE" -> dec = ","

If pp_lang = "EN" -> dec = "."

Value

data.frame (Invisible)

Examples

data("cardano")
parse_cardano(cardano, pp_lang = "DE")
#> ℹ Setting decimal separator to `,` because `pp_lang = 'DE'` is set. If this is
#> undesired and/or you want to silence this warning, please explicitly set
#> argument `dec`.
parse_cardano(cardano, pp_lang = "EN")
#> ℹ Setting decimal separator to `.` because `pp_lang = 'EN'` is set. If this is
#> undesired and/or you want to silence this warning, please explicitly set
#> argument `dec`.