Parse Polkadot Staking Rewards

Retrieves staking rewards for a specific account from https://subscan.io.

parse_polkadot(
  address,
  pp_security_name = "Polkadot",
  currency = "EUR",
  pp_lang = "EN",
  securities_account = NULL,
  dec = NULL,
  sep = ";",
  filename = NULL,
  api_key = NULL
)

Arguments

address
[character]
Wallet address.

pp_security_name
character
The name of the security in Portfolio Performance (e.g. "Kusama").

currency
character
Currency in which to calculate the staking rewards. Must be part of a valid coin pair on Binance, e.g. "DOTEUR".

pp_lang
character
The currency of Portfolio Performance. Only German ("DE") and English ("EN") are supported.

securities_account
character
The name of the account to which to associate the deliveries. In German called "Depot".

dec
character
Decimal separator. If not set, inferred by the value of argument pp_lang:

If pp_lang = "DE" -> dec = ","

If pp_lang = "EN" -> dec = "."

sep
character
Columns separator. Defaults to ;.

filename
character
The output filename.

api_key
character
Optional personal API key.

Value

tibble::tibble (invisibly)

Examples

parse_polkadot("1qEk2g6N1uugFgyvmnsL6P9Conh5nTwL5mj85bm3XHtjc73", dec = ".")
parse_polkadot("1qEk2g6N1uugFgyvmnsL6P9Conh5nTwL5mj85bm3XHtjc73",
  api_key = Sys.getenv("subscan_api_key"), dec = "."
)